Add arduino interface capability
Modified the run_sensor.py file to receive serial data from the arduino.
Created an arduino folder in the project root that contains the sketch
that is running on the arduino. The sketch records time at microsecond
resolution when a rising edge is detected. It outputs an array that
contains the difference of two consecutive array positions resulting in
the time duration between two rising edges.