Edit db_interaction
Edited the name of the variables that the user_insert function and the
textfile_insert function are assigned to. The value of the variable is
the row_id number for that newly inserted row. This might not work for
the user table because the table doesn't have an autogenerated integer
primary key.