Ods file integration
The pyoo code to use openlibre office is working when an instance of
libre office is already running. Still having issues with calling an
instance from python code. The command to start libreoffice is done
manually. Importing the data from a textfile to libreoffice is working
only when it's one element from the lines array into one cell in
libreoffice. Still tying to work out haow to populate the entire column
with the contents of the textfile.