Add reload function to reload new values after rerun the program
to re-run the program again without restrating, we need to call reload
function to reload new values that need to be displayed in the final results page. but it
doesn't seem to work  properly. it's working right now, but it takes
calling the run function twice to reload new values, and also,it effects the look of the results page. So,I think there is no need to use this functionality