Graph finished
The graphing function is finally done. The first value in the torque
array is ommitted when building the graph. This is the only way I was
able to get the graph to look like the one in the spreadsheet. Also the
torque y axis might need to be adjusted later.