Fix multiple run scenario bug
There was an issue when trying to run the program with the run again and
new user checkboxes on the results page. The problem wasn't in the GUI.
It was in power.py. The lines list that holds the sensor data wasn't
clearing. New sprint data was being appened to that list. Fixed by
clearing the lines list after it's use in the power_sheet function.